12_english_ls/requirements.txt
langchain
python-dotenv
langchain_community
openai
streamlit